Change margin and padding values for exercice
I didn't directly understand the solution for the thrid block to be
pushed to the right inside the 600px body. First, I juste added more
pixels to, after some tries, match the desired_outcome. But the solution
uses "margin-left: auto;" which I didn't understand at first. Indeed,
the third box has a specified width and so if you add the property of
auto on a margin it adjust itself between the space available inside the
parent.