Complete exercice 1 but with twist
I know I could have completed it with only container properties like
in the solution: "justify-content: center;" and "align-items: center".
But I remembered when you have only one item, you can simply margin=
auto and the margins will take the space available and so center the
element. I did it the smart way but not the exercice way. Cheers !