Complete exercice two with too much complexity
Indeed I once again did too much until I realise it isn't necessary
to use the margin: auto on the logo and applying flex: auto
on every element. Because justify-content: space-between does exactly
the same job to flex items that can't gro because the default flex state
is 0 1 auto. I know it's how we learn but I think it reflects the
complexity of the flex layout.